+++ .prettierrc
@@ -1,6 +1,3 @@
 {
-  "plugins": ["prettier-plugin-tailwindcss"],
-  "prettier/prettier": ["error", {
-      "endOfLine": "auto" }
-  ]
+  "plugins": ["prettier-plugin-tailwindcss"]
 }
