+++ src/components/business/record/index.tsx
@@ -3,9 +3,9 @@ import React from 'react'
 import { getDictionary } from '@/dictionaries/get-dictionary'
 import { LocaleEnum } from '@/types/locales'
 
-import AudioRecorder from './audioRecord'
+import AudioRecorder from './audio-recorder'
 
-async function RecordPage() {
+async function RecordComponent() {
   const dictionary = await getDictionary(LocaleEnum.EN)
 
   return (
@@ -15,4 +15,4 @@ async function RecordPage() {
   )
 }
 
-export default RecordPage
+export default RecordComponent
