+++ src/dictionaries/en.json
@@ -1,74 +1,11 @@
 
 {
   "Contact Us": "Contact Us",
-  "About us": "About us",
-  "Service": "Service",
-  "Outsource": "Outsource",
-  "Product": "Product",
-  "Investment": "Investment",
-  "Assistance": "Assistance",
-  "Team": "Team",
-  "Menu": "Menu",
-  "FAQs": "FAQs",
-  "Privacy & Policy": "Privacy & Policy",
-  "© 2077 Untitled UI. All rights reserved.": "© 2077 Untitled UI. All rights reserved.",
-  "About Us": "About Us",
-  "News": "News",
-  "Newsletter": "Newsletter",
-  "Email Address": "Email Address",
-  "When Xclusive still exists, Xclusive products will still exist, and Xclusive clients will not be left behind": "When Xclusive still exists, Xclusive products will still exist, and Xclusive clients will not be left behind",
-  "Get Started": "Get Started",
-  "XCLUSIVE SERVICE": "XCLUSIVE SERVICE",
-  "Multiple Programming Languages & Framework": "Multiple Programming Languages & Framework",
-  "Xclusive Daily Activities": "Xclusive Daily Activities",
-  "Check out our latest news": "Check out our latest news",
-  "Happy Clients": "Happy Clients",
-  "Completed Projects": "Completed Projects",
-  "Hours Of Support": "Hours Of Support",
-  "Roadmap background": "Roadmap background",
-  "Roadmap": "Roadmap",
-  "Xclusive Skills": "Xclusive Skills",
-  "FAQ's": "FAQ's",
-  "Expand question": "Expand question",
-  "Let's Discuss your Projects": "Let's Discuss your Projects",
-  "We pride ourselves with our ability to perform and deliver results. Use the form below to discuss your project needs with our team, we will get back asap": "We pride ourselves with our ability to perform and deliver results. Use the form below to discuss your project needs with our team, we will get back asap",
-  "Xclusive Products": "Xclusive Products",
-  "View previous team member": "View previous team member",
-  "View next team member": "View next team member",
-  "Website, Application, Dapp, Game,...": "Website, Application, Dapp, Game,...",
-  "XC token, XC products, XC hotels, XC Real Estate.": "XC token, XC products, XC hotels, XC Real Estate.",
-  "Business consulting, technical assistance": "Business consulting, technical assistance",
-  "USDT P2P Xclusive": "USDT P2P Xclusive",
-  "USDT P2P Lending Xclusive": "USDT P2P Lending Xclusive",
-  "Support for failed projects": "Support for failed projects",
-  "Arbitrage bot": "Arbitrage bot",
-  "Domain & Hosting": "Domain & Hosting",
-  "Dark money-fueled law firm joins Massachusetts climate suit against Big Oil": "Dark money-fueled law firm joins Massachusetts climate suit against Big Oil",
-  "Law firm conflicts 'permeated FTX's bankruptcy', professors allege": "Law firm conflicts 'permeated FTX's bankruptcy', professors allege",
-  "Freshfields the Latest to Move to 4 Days in US Offices": "Freshfields the Latest to Move to 4 Days in US Offices",
-  "Develop outsource": "Develop outsource",
-  "Product for Korea": "Product for Korea",
-  "Product for Japan": "Product for Japan",
-  "Establishment of a joint stock company between Vietnam and Japan outsource, product": "Establishment of a joint stock company between Vietnam and Japan outsource, product",
-  "Establishment of a joint stock company between Vietnam and Korea for outsource, product": "Establishment of a joint stock company between Vietnam and Korea for outsource, product",
-  "30%: outsource development": "30%: outsource development",
-  "70%: focus on internal products of Japan-": "70%: focus on internal products of Japan-",
-  "Vietnam and Korea-Vietnam": "Vietnam and Korea-Vietnam",
-  "Listing XC token on BTC Network": "Listing XC token on BTC Network",
-  "Chief Executive Officer Project Manager Technical Leader": "Chief Executive Officer Project Manager Technical Leader",
-  "UI/UX Designer": "UI/UX Designer",
-  "Address": "Address",
-  "Phone": "Phone",
-  "Email": "Email",
-  "Full name": "Full name",
-  "Enter your name*": "Enter your name*",
-  "Enter your email*": "Enter your email*",
-  "Subject": "Subject",
-  "Enter subject": "Enter subject",
-  "Messenger": "Messenger",
-  "Write your question or suggestion...": "Write your question or suggestion...",
-  "Send Messenger": "Send Messenger",
   "Your browser does not support the audio element." : "Your browser does not support the audio element.",
   "Download Recording": "Download Recording",
-  "hello": "hello"
+  "hello": "hello",
+  "Stop Recording": "Stop Recording",
+  "Start Recording": "Start Recording",
+  "Audio Recorder": "Audio Recorder",
+  "Captions": "Captions"
 }
